Refactor orderbook, Redis, websocket & handlers
Code cleanup and small refactors across order book, Redis persistence, websocket server/handlers, and various handlers/types. Key changes:

- Converted several ToString impls to std::fmt::Display in order_book::db_persistence and replaced or_insert_with(VecDeque::new) with or_default().
- Simplified order matching flow in engine (unified trades assignment) and improved FOK handling.
- Redis persistence: simplified price key formatting, use next_back() for parsing, added explicit generic annotations for redis commands, marked several methods with #[allow(dead_code)].
- WebSocket server/handlers: derive Default for WebSocketServer, removed manual Default impl, simplified send/do_send calls, avoided unused warnings by prefixing or renaming vars, added some #[allow(dead_code)] and removed duplicated Default impls.
- Handlers and utils: removed unused imports, used strip_prefix for subscription parsing, initiated MarketMakerConfig via struct literal, and added various lint allowances (dead_code, clippy) in types and utils.

These changes are mostly cosmetic and lint-driven to reduce warnings and improve code clarity without changing core behavior.

+++ src/handlers/transaction_handler.rs
@@ -7,7 +7,6 @@ use actix_web::{web, Error, HttpResponse, Result};
 use deadpool_redis::Pool;
 use entity::{transaction, users};
 use rust_decimal::Decimal as RustDecimal;
-use sea_orm::prelude::Decimal;
 use sea_orm::{
     ActiveModelTrait, ColumnTrait, DatabaseConnection, EntityTrait, PaginatorTrait, QueryFilter,
     QueryOrder, QuerySelect, Set, TransactionTrait,
@@ -77,7 +76,7 @@ pub async fn deposit_money(
     // Update user balance
     let mut user_active_model: users::ActiveModel = user.into();
     user_active_model.wallet_balance =
-        Set(Decimal::from(RustDecimal::try_from(balance_after).unwrap()));
+        Set(RustDecimal::try_from(balance_after).unwrap());
     user_active_model.updated_at = Set(chrono::Utc::now().naive_utc());
 
     let _updated_user = user_active_model.update(&txn).await.map_err(|e| {
@@ -90,11 +89,9 @@ pub async fn deposit_money(
     let transaction = transaction::ActiveModel {
         user_id: Set(user_id),
         r#type: Set("deposit".to_string()),
-        amount: Set(Decimal::from(RustDecimal::try_from(amount).unwrap())),
-        balance_before: Set(Decimal::from(
-            RustDecimal::try_from(balance_before).unwrap(),
-        )),
-        balance_after: Set(Decimal::from(RustDecimal::try_from(balance_after).unwrap())),
+        amount: Set(RustDecimal::try_from(amount).unwrap()),
+        balance_before: Set(RustDecimal::try_from(balance_before).unwrap()),
+        balance_after: Set(RustDecimal::try_from(balance_after).unwrap()),
         status: Set("completed".to_string()),
         reference_id: Set(reference_id.clone()),
         created_at: Set(chrono::Utc::now().naive_utc()),
@@ -223,7 +220,7 @@ pub async fn withdraw_money(
     // Update user balance
     let mut user_active_model: users::ActiveModel = user.into();
     user_active_model.wallet_balance =
-        Set(Decimal::from(RustDecimal::try_from(balance_after).unwrap()));
+        Set(RustDecimal::try_from(balance_after).unwrap());
     user_active_model.updated_at = Set(chrono::Utc::now().naive_utc());
 
     let _updated_user = user_active_model.update(&txn).await.map_err(|e| {
@@ -236,11 +233,9 @@ pub async fn withdraw_money(
     let transaction = transaction::ActiveModel {
         user_id: Set(user_id),
         r#type: Set("withdraw".to_string()),
-        amount: Set(Decimal::from(RustDecimal::try_from(amount).unwrap())),
-        balance_before: Set(Decimal::from(
-            RustDecimal::try_from(balance_before).unwrap(),
-        )),
-        balance_after: Set(Decimal::from(RustDecimal::try_from(balance_after).unwrap())),
+        amount: Set(RustDecimal::try_from(amount).unwrap()),
+        balance_before: Set(RustDecimal::try_from(balance_before).unwrap()),
+        balance_after: Set(RustDecimal::try_from(balance_after).unwrap()),
         status: Set("completed".to_string()),
         reference_id: Set(reference_id.clone()),
         created_at: Set(chrono::Utc::now().naive_utc()),
